Non-forest fixes: is_forest was not being reset between pixels, causing weird interactions when non-forest and forest pixels were mixed in a simulation.
Changed the conditions under which spinup is skipped - now only skipped for completely non-forest pixels that have no growth increments. Spinup should not necessarily be skipped for pixels where an initial pool value has been specified - some users would like to be able to specify a value for the slow pool in forested pixels, in which case spinup still needs to run to initialize the rest of the pools.